+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [0.6.0] - ongoing work
 ### Added
+- Hyperedge.atom and .not_atom properties.
 - Functional patterns: var, atoms, lemma.
 - Moved all patterns functions to graphbrain.patterns.
 - Hypergraph.text().